backend/requirements.txt
fastapi==0.115.8
uvicorn[standard]==0.35.0
tensorflow==2.20.0
pydantic==2.11.1
python-multipart==0.0.20
python-dotenv==1.1.1
